Fixes a bug with showing the Geolocation permissions prompt
Currently, if an iframe causes the Geolocation permissions prompt to be shown,
but the frame is navigated away before the user responds to the permissions
prompt, the prompt is not removed from the screen.

This change fixes that bug. It uses the new
ChromeClient::cancelGeolocationPermissionRequestForFrame() method, which was
cherry-picked from WebKit in https://android-git.corp.google.com/g/#change,41747

The logic is complicated by the fact that multiple iframes may have requested
Geolocation permissions for the same origin. In this case, we should not hide
the prompt until the last such frame is navigated away.

Bug: 2463551
Change-Id: I1e4de05586150c7a94bc8343f6e56e4a4870cd3b