Fixes a bug in RenderLayer::paintLayer due to Android's FASTER_MATRIX optimisation
This bug was introduced by the merge to r51976 on 15 Dec 2009.

WebKit change 51355 modified RenderLayer::paintLayer and the Android
FASTER_MATRIX modification was not updated to reflect this. The difference is
that, compared to the intended code, the Android version misses a call to
makeMatrixRenderable, which in turn calls makeAffine, when calculating the
transform matrix.

See http://trac.webkit.org/changeset/51355

Change-Id: Ibc25b92e64381642fc3bbcfd672c056a3d89e035