Fixes a crash in the Geolocation service when using maximumAge
When a browser page is put to the background, the Geolocation object is
suspended. If the Geolocation object has active requests, it suspends its
Geolocation service.

Now that we have maximum age support, the Geolocation object can have active
requests without having started its Geolocation service. In this case, the
geolocation service can ignore calls to suspend and resume.

This avoids a crash when the Android Geolocation service hasn't created it's
java bridge when suspend is called.

Bug: 2453611
Change-Id: I368226281920512584a92b5f1967d602df1f0782