Adds layout tests for HTML5 features
The following layout tests should all pass on Android, as they are for recently
implemented HTML5 features ...
- fast/dom/Geolocation
- storage
- http/tests/appcache

This change adds these tests to the Android tree, at the current WebKit revision
r54731. This is so that we can easily keep track of which tests should always be
green, and so that we can add Android-specific test results.

We also add the following paths ...
- fast/js/resources - used by the Geolocation tests
- http/conf - used by the Appcache tests

Tests that are currently failing are added to the DumpRenderTree skipped list
temporarily, to keep all tests green.

Change-Id: Id96c05e3746ed64e4e4c40c99567b8def688f90a