fixing problem were plugin surfaces are not created.
The problem occurs when a page containing a plugin is refreshed
or the plugin is loaded using javascript. If the window size
is set before the plugin view has its parent sent then the
plugin surfaces were not being initialized. This change ensures
that if that happens the surfaces will still be created.

see bug #2476059.