Serialize/unserialize screen width scale, which is used when restoring sites.
Also change the scales' default value to 0, so in WebViewCore.java, in case
the scales are not set inside webkit, the default logic can be used.
This helps in case a mobile site is loaded but no scales is saved, and when
restored, the mobile site will have wrong restored scales because default
values are not 0.