adjust cache layer position by offsets
Nodes in layers need to know where they are relative to the layers
so that the cursor rings are correctly positioned as the layer
moves (in document coordinates). In addition to tracking the
global offset to make the coordinates relative to any parent
frames, the node is offset relative to the body.

Sometimes layers have zero height or width; in this case, don't
treat them as clips.

Change-Id: Id2811c31a4a0674d316aadda210570ec93311013
http://b/2503096