Cherry-pick WebKit change 54972 to merge DOMWindow and WorkerContext object wrapping in V8 bindings
The motivation for this change is to disable workers on Android, because V8 on
Android does not have the required locking.

The current version of WebKit in Android (54731) uses workers features without
ENABLE(WORKERS) guards in the V8 bindings. In particular,
V8DOMWrapper::instantiateV8ObjectInWorkerContext is used from generated code
without guards.

Fixing the code generator would be difficult and would be wasted effort as
V8DOMWrapper::instantiateV8ObjectInWorkerContext no longer exists in tip-of-tree
WebKit. Instead, we cherry-pick the WebKit change which removes this method.

See http://trac.webkit.org/changeset/54972

Change-Id: Ie6365073d7a4d92aa7c1553d87a7e1c1ed514118