Initial check-in of the external/webkit part of the chrome http stack.
There are no changes in WebCore anymore, the new files live in WebKit/android/WebCoreSupport. There are two WebCore functions implemented in WebUrlLoader.cpp, they are
one-liners that call a WebKit function or return a bool.

Only async downloads are implemented so far. There are crashes on some websites, but better to get things checked in, and then start fixing outstanding issues.

I have tried to address all the issues from the reviews, but since the code has changed a bit I have not replied to each of them.

Change-Id: Ie83798e2f82ce78cb00ee5a4a558c9d8085c9567