cherry-pick 74f79778892548221468d075f949acb3d4064368 to froyo
In Android, as we always remap the plugin from page
space to view space, we do want to update when the rect
in the page space changed. This should fix the bug where
the plugin is inside a fixed position. So even its
page coordinate changed when scrolling, its window
coordinate didn't change. We missed the update.

Fix http://b/issue?id=2542934