don't hide transparent nodes
A node may be transparent if the body of the node is drawn earlier.
In this case, the node may not be tested to see if it has been
occluded by later drawing, since no drawing inside the scope of
the node is actually visible. So, skip the hidden test for
transparent nodes.

Change-Id: Ib748e9e7b86252f791ee68198d1d794fb4591a88
http://b/2582455