Merge webkit.org at r58033 : Move Android iframe-flattening code to RenderIFrame.cpp
The Android iframe-flattening code was in RenderPartObject. The rendering
classes have been refactored upstream in http://trac.webkit.org/changeset/57866
so that most of RenderPartObject was moved to RenderIFrame. RenderPartObject was
then removed in http://trac.webkit.org/changeset/57900.

Iframe-flattening was also added upstream in
http://trac.webkit.org/changeset/56718.

This change takes the Android iframe-flattening code from RenderPartObject and
moves it to RenderIFrame. It also renames FLATTEN_IFRAME to
ANDROID_FLATTEN_IFRAME and FLATTEN_FRAMESET to ANDROID_FLATTEN_FRAMESET for
clarity.

Change-Id: I1af8d042f8d0c4abd513f8a315ac681360c1cc53