Enable contentEditable.
WebViewCore.cpp:
Open the keyboard when a contentEditable element is put
into focus by a click.
In key(), return whether the selection changed if a contentEditable
element is in focus.

CacheBuilder.cpp:
Add root contentEditable elements to the navigation tree.

CachedNode.h:
Include contentEditable elements as elements that want key events.

CachedNodeType.h
Add a type for contentEditable.

WebView.cpp:
Do not call setFollowedLink for contentEditable, so that the orange
selection ring stays around the field.
Add a check to determine whether the page should handle shift and
arrow keys.

Bug 1788820

Caveats:
Does not ensure that the caret remains on screen.  Frame::revealSelection
is called, but we ignore it for other reasons.  Need to investigate that.
The cursor will blink if the contentEditable node has focus, even if the
user has not clicked on it or has moved to a different input field. Further,
while in this state, the user can input text.

Requires a change to frameworks/base

Change-Id: Ife39254f46dcc1046a075eee2fda6cf4879b4ee8