Update Android to reflect upstreamed code to avoid Geolocation leaks
This is a cherry-pick of WebKit change http://trac.webkit.org/changeset/60069
This releases references to JS callbacks in the page and allows the Frame and
thus the Geolocation object to be cleaned up. This change allows us to remove
the Android-specific code which achieved the same through an unload listener.

Bug: 2169518
Change-Id: I1ed97d40d7f69ec94a58eaa21bfa4979737cc84d