No longer consider a semi-colon to be a separator character when parsing the meta-tag
The Android meta-tag parsing code was originally copied from the iPhone
implementation. A later change added the semi-colon to the list of separator
characters, motivated by Reader. See http://b/issue?id=1416841 and
https://mondrian.corp.google.com/changelist/115656-p9

Meta-tag support has since been added to WebKit in
http://trac.webkit.org/changeset/57775. The upstream parsing code does not
consider the semi-colon to be a separator. The upstream parsing code was merged
into Android as part of the merge to WebKit r58033 in
https://android-git.corp.google.com/g/#change,50648

This change removes the Android-specific changes so the semi-colon is no longer
considered a separator. This brings Android back into line with iPhone and other
browsers.

Change-Id: Id4cbbadb414d0508a89006d9beb91c56fa1f8daf