don't use animated gifs if gifs only have one frame
Gifs use a different decoding mechanism from other images. The
default webkit implementation allocates a bitmap for the gif,
which is serialized when the picture is flattened. A page with
a 25K display list may flatten into a 20M file if GIFs are
included as bitmaps.

GIFs can be written like other image types, as streams; then
the display list flattens to a 140K file. Currently, the
Skia-based images only can handle non-animated types.

When animated gifs was recently enabled, the flattened file
size grew. All gifs were treated as if they could be animated,
and all were decoded into bitmaps with straight memory allocations.

This change uses the stream-allocation mechanism if the GIF
has only one frame, and is therefore not animating.

Change-Id: Ie2360bb6f20515b81bd62761ca97242f33ffb43b
http://b/2650945