ignore offscreen drawing in SVG
Drawing outside of the visible window is recorded so that future
scrolls have the correct picture data. When SVG triggers this
code path on its way to drawing to an offscreen bitmap, it
has not instantiated a WebViewCore object. Since in SVG's case,
there's no reason to record drawing outside if the offscreen
bitmap, it's correct to do nothing.

Offscreen bitmaps do not work with SVG; but that problem is
independent of this bug fix.

Change-Id: I0b9c6d5506b6a35b2992abfa6bcb1db6cb872fd0
http://b/2682606