First draft to add a fat point to WebKit hit test.
If padding is 0, it is the old style point hit test.
If it finds a node, it will break the search loop
and return.

If padding is non-zero, the first node will be added
to HitTestResult's innerNode as before. But instead
of finishing the search, hit test will continue to
look for other nodes covered by the fat point unless
it is fully inside the current node.

Here are some highlights of the changes.

. instead of testing renderRect.contains(x, y),
  we now test renderRect.intersect(pointRect)
. when a Node is hit, it will be appended to the
  rawNodeList of HitTestResult. So the order inside
  the rawNodeList represents the hit test order.


Currently the fat point doesn't support SVG nodes.
It also doesn't support overflow control yet.