reverse RTL characters in copy
Text is copied from the picture in visual left-to-right order.
Reverse right-to-left substrings before returning the result.

This algorithm for reversal is also used by webkit to draw
text, and has the same limitations.

This fixes text pasted into the title bar, but does not work
correctly for text pasted into an input field -- LTR text is
placed incorrectly -- but that's another bug.

Change-Id: I4709b74e32495d4b77d33910ac8da34d4a88edd0
http://b/2525085