Refix http://b/issue?id=2509597.
Instead of changing the screenRect for the Screen
object, change createWindow() logic to depend on the
WebCoreViewBridge width/height instead of Screen
as Screen is supposed to be in device pixel and it
is not proper to be used here. Update the comment
for the screenRect so that we won't mistakenly modify it.

Fix http://b/issue?id=2690479