Fix for b/2786464
Some renderobjects are "anonymous", i.e. have the document node associated with them and in this case
the node() function returns null. We were then using the null RefPtr, hence the crash. RenderListMarker
is an example of such an anonymous node which is why websites that use lists exposed the crash.

Change-Id: I1a5527c42d99bbe8418b0e7bce4f3c245276e058