set maximum image size
Our use of web settings does not specify the maximum image size,
although we have some hard-coded values used when decoding the
image. For defective images, the hard-coded values are not enough,
since we may exceed memory in the decoder before the limit is
reached.

Pass along our limits to the decoder, and respect any limits
set by the caller.

companion change in frameworks/base

Change-Id: Icdfca87a44a3af8bcc1f0ed4792e04e060a67333
http://b/2689053