refine text selection
Change 'select all' to begin with the first character and end
with the last rather than relying on the top left and bottom
right. The old algorithm failed for pages broken into multiple
columns.

Similarly, capture the text from the beginning selection to the
ending selection, rather than relying on the characters contained
by the selection region.

Reduce the gap required to add a linebreak from twice the
lineheight to 1.5x the lineheight.

Change-Id: Ib90a719ea3353feeaa3a6cc692d0dd9cff498cb5
http://b/262451