Merge WebKit at r61871: Fix conflicts in HTMLDocumentParser.cpp
New document parser in webkit. Just taking in the new parser.
See http://trac.webkit.org/changeset/61674

This file is now what is in webkit at the time of the merge, with ANDROID_INSTRUMENT added.

The instrumentation for Parsing now includes inline script execution, this will have to be fixed later. Updated the
description for the parsing counter.

Change-Id: I2c1bae725fe8989999620430e4745e1dfa7c450b