Fix for bug 2816268: Refactor ANDROID_USER_GESTURE code
Used the UserGestureIndicator class instead of the local ANDROID_USER_GESTURE code. I have made a simple test case for the problem at:
http://updates.monsen.it/source.html
This cannot be made into a layout test since the Android Dumprendertree does not support focus on windows other than the main one.

Lots of local changes removed, but not many actual changes. Most are in parameters sent to other functions.

As far as I can see they user gestures are only checked for in ChromeClientAndroid.cpp and WebCoreFrameBridge.cpp, and I
replaced ANDROID_USER_GESTURE with UserGestureIndicator::processingUserGesture().

The UserGestureIndicator class was added to webkit in this CL:
http://trac.webkit.org/changeset/57045

ResourceRequest.h is a cherry-pick of this CL from webkit.org
http://trac.webkit.org/changeset/62668

Change-Id: Ia254b2cee833459fcc2cacfb2751ef7a29a8f64a