Enable composite layers for the sub frame.
Fix the visibleContentRect for iframe. It should be
relative to its parent instead of the viewport.

For fixed position, we still check for the top frame
as the current logic positions the object relative to
the screen which only applied to the top frame.

For plugin, it depends on PluginWidget's platformLayer().
The default is 0.

For video, it depends on MediaPlayer's supportsAcceleratedRendering().
The default is false.

Fix the crash in the GraphicsLayerAndroid. The root
layer is a container layer, there is no need to draw
them. As LayerAndroid doesn't create Picture for them,
we should ensure to skip draw.

Fix http://b/issue?id=2733947