handle out of order selections
Usually, the text on a webpage is drawn from top to bottom but
sometimes it is drawn out of order. In these cases the picture
finds the select end before it finds the select start.

When the select start and end are ordered, all of the text in
the picture between the two are selected. This allows selecting
all of the text on the page, even if the start isn't spacially
before the end.

When the select start and end are flipped, select the text
spacially between the two.

The flipped page scenario was found by
http://code.google.com/p/android/issues/detail?id=8533

Change-Id: Ia22d49396d68fefc3d64fad438c4b0c02f668c1c