Combine base PictureSet with composite layers into one SkLayer tree. The BaseLayerAndroid is the root of the composite layers and it has the PictureSet as its base content.
Before, WebViewCore uses m_contentMutex and both UI
and WebCore threads access the m_content. Now we use
the layer approach. When WebCore thread updates its
content in recordContent, it creates a new BaseLayerAndroid
and copy both PictureSet and composite layers into it.
Then it is sent to be consumed by UI thread.

Clean up sync layer in both ChromeClientAndroid and
GraphicsLayerAndroid.

splitContent can be a little tricky with this change.
Now UI has its own copy of PictureSet. When it takes
too long to draw, it will send a request to WebCore
to split the PictureSet. When it is done, a copy of
the new PictureSet will be sent back to UI.

There is a matching change in framework/base