Add dom/xhtml/level2 layout tests at WebKit r62496
These all pass.

Note that the Android-V8-specific results for the HTMLDocument tests are
required because the tests rely on Error.line, which is non-standard and not
provided by V8.

The Android-specific result for HTMLFormElement10.xhtml is required because the
test appears to be flawed, and navigates the page before notifyDone is called.

None of these tests are run on Chromium.

Change-Id: I491d87a9dac4c132093369d1bde974baed77a757