Add dom/xhtml/level3 layout tests at WebKit r63173
These all pass.

Note that the Android-V8-specific results are required because these tests rely
on Error.line, which is non-standard and not provided by V8.

A further 15 tests have no expected results so will be added to the
DumpRenderTree ignore list.

None of these tests are run on Chromium.

Change-Id: I6b7ea492586fa2d9cd60858f06641941c87ede14