Enable scrollable divs.
Force a composite layer when the style says the content is scrollable.
Record the border and background in the main content picture.  When
the contents of the layer are bigger than the size, record the
foreground contents in a separate picture which is clipped by the
border and size.

When updating the base layer, remember the scroll position of each
layer and update the new layer with the same position.

Bug: 1566791
Change-Id: If440e4f215db6bda9df32a781d754d1f5a238162