select text out of order
When the selected text visual order doesn't match the picture
order, the text is selected spacially. The rectangle described
by the start and end points limits what text is selected.

This can fail when the rectangle described is too narrow to
enclose all the lines between the top and bottom. This change
extends the lines by including the text adjacent to the start
and end when computing the limit bounds.

And:

- Refactor the code so that drawing the region and selecting
the text can share this logic.

- Add slashes as characters that prevent inserting spaces at
the ends of lines. (ASCII characters space, dash, slash, and
backslash cause lines to wrap.)

- Narrow the error term for detecting spaces. The 1/2 value
before inserted spaces incorrectly.

Change-Id: I645f38dc246c61b1bc7c94e61553e5e6e36e3f85
http://b/2817635