Removing the list of active loaders
A loader will not be deleted before the WebRequest has called didFinishLoading() now, so there is no longer a need for this.
There was also a problem with that list. If a loader got deleted and a new one created with the same pointer value the deleted one could be mistaken for the new one.

Change-Id: I856519e751f6f1d15cfbd426a2cd2ba71315850b