Move check to clear the text input to the webcore thread.
Originally written in response to bug 2835685, which was
fixed in
https://android-git.corp.google.com/g/#change,61059

WebViewCore.cpp:
Send a message to clear the text input/remove the soft
keyboard if the click results in no text input field
being in focus.

WebView.cpp:
Remove the old checks to clear the WebTextView, which
is now done in WebViewCore.cpp.  This makes clearing it
consistent with the checks for opening the WebTextView.

Change-Id: I06ba7aadc95302c90af67da16edccd46896fcdbf