Update navigation in scrollable layers.
Set the foreground clip after drawing.  Use the absolute bounds to
compute the local foreground clip in order to compensate for any
outline.

Consolidate the check for overflow scrolling into RenderLayer.
Request a compositing update after computing the scroll dimensions.
Only change the foregroundRect of the layer during paint so that the
outline rect (and background/layerBounds) are correct.

Draw the outline as part of the background phase.  During painting of
a layer, scroll to (0,0), paint, then scroll back.

When clicking on an element in a layer, scroll to the position of the
element but do not scroll back.  This makes text input fields visible
to the tree and will properly update when typing.  Record the original
scroll position of layers in order to offset the bounds of nodes when
checking the nav cache.  Make sure to reset all cached layers during
setRootLayer.  Otherwise we were reaching into layers from the wrong
thread.

Change-Id: Id9827ec461989b0869a8252d4d2563ecd12c5fac